Recommit Joerg Sonnenberger patch with more fix.
When writing jpeg scanline, it wouldn't fail with FILE* as an output, but will
with a data buffer. So we need to re run the jpeg routine on this scanline. As I
don't know if the buffer is always preserved during two calls to the jpeg
library, I am copying it again.



SVN revision: 35804